feat(ny): Fix NY Court of Appeals
Currently we are being blocked -
so we need to resolve this issue for
such an important court

the downside is no PDFs for now
